ENH: add capability to init from existing segmentations
Added an attribute on the finding node to pass along the id of a
prior segemented label map volume to use when adding a finding.
The module workflow is split up across multiple small functions
so the passing of the prior volume information isn't as clean
as it could be with a redesign.

Issue #1